Align RemoteCommand + RemoteCommand.Response + RemoteCommandRequest with Android-Java
