+++ remotecommanddispatcher/src/main/java/com/tealium/remotecommanddispatcher/RemoteCommandConstants.kt
@@ -1,10 +1,5 @@
 package com.tealium.remotecommanddispatcher
 
-enum class RemoteCommandType(val type: String) {
-    WEBVIEW("webview"),
-    JSON("json")
-}
-
 object Settings {
     const val DLE_PREFIX = "https://tags.tiqcdn.com/dle"
