+++ remotecommanddispatcher/src/main/java/com/tealium/remotecommanddispatcher/RemoteCommandRequest.kt
@@ -12,13 +12,13 @@ import java.util.regex.Pattern
  * The object used to request a remote command to the remote command dispatcher.
  * A response object must be set in order for the command to be invoked.
  */
-data class RemoteCommandRequest(var commandId: String? = null,
-                                var payload: JSONObject = JSONObject(), // comes from tiq or payload of track call when using json rc
-                                var response: Response? = null) {
+class RemoteCommandRequest(var commandId: String? = null,
+                           var payload: JSONObject = JSONObject(), // comes from tiq or payload of track call when using json rc
+                           var response: RemoteCommand.Response? = null) {
 
     companion object {
         fun jsonRequest(remoteCommand: RemoteCommand, mappedPayload: JSONObject): RemoteCommandRequest {
-            return RemoteCommandRequest(remoteCommand.commandId, mappedPayload, Response(remoteCommand.commandId, requestPayload = mappedPayload))
+            return RemoteCommandRequest(remoteCommand.commandId, mappedPayload, RemoteCommand.Response(remoteCommand.commandId, requestPayload = mappedPayload))
         }
 
         fun tagManagementRequest(request: String): RemoteCommandRequest? {
@@ -68,8 +68,8 @@ data class RemoteCommandRequest(var commandId: String? = null,
             return remoteCommandRequest
         }
 
-        private fun createResponse(commandName: String, id: String?, args: JSONObject): Response {
-            return object : Response(commandName, id, args) {
+        private fun createResponse(commandName: String, responseId: String?, requestPayload: JSONObject): RemoteCommand.Response {
+            return object : RemoteCommand.Response(commandName, responseId, requestPayload) {
                 override fun send() {
                     responseId?.let {
                         body?.let {
@@ -102,4 +102,6 @@ data class RemoteCommandRequest(var commandId: String? = null,
 
         private val PROTOCOL_PREFIX = Pattern.compile("^$TEALIUM_PREFIX.+", Pattern.CASE_INSENSITIVE);
     }
+
+
 }
