+++ remotecommanddispatcher/src/main/java/com/tealium/remotecommanddispatcher/TealiumConfigRemoteCommandDispatcher.kt
@@ -2,7 +2,6 @@ package com.tealium.remotecommanddispatcher
 
 import com.tealium.core.TealiumConfig
 
-const val REMOTE_COMMAND_ENABLED = "remote_command_enabled"
 const val REMOTE_COMMAND_CONFIG_REFRESH = "remote_command_config_refresh"
 
 /**
@@ -15,18 +14,5 @@ var TealiumConfig.remoteCommandConfigRefresh: Long?
     set(value) {
         value?.let {
             options[REMOTE_COMMAND_CONFIG_REFRESH] = it
-        }
-    }
-
-/**
- * Optional config flag for the remote command dispatcher.
- *
- * Set this key to true in TealiumConfig to use TiQ for remote commands.
- */
-var TealiumConfig.remoteCommandEnabled: Boolean?
-    get() = options[REMOTE_COMMAND_ENABLED] as? Boolean
-    set(value) {
-        value?.let {
-            options[REMOTE_COMMAND_ENABLED] = it
         }
     }
