+++ remotecommanddispatcher/src/main/java/com/tealium/remotecommanddispatcher/remotecommands/HttpRemoteCommand.kt
@@ -4,7 +4,6 @@ import android.net.Uri
 import com.tealium.core.Logger
 import com.tealium.core.network.*
 import com.tealium.remotecommanddispatcher.BuildConfig
-import com.tealium.remotecommanddispatcher.Response
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.runBlocking
